Release 0.5.0 (#13)
* chore: upgrade version to 0.5.0

* feat: create factories for all use cases (#14)

* User authenticate controller (#15)

* chore: install zod package

* feat: update error handling to catch zod error instances

* feat: create env file with zod validation

* feat: update prisma seed

* feat: update env var from prisma instance

* feat: add and update user authentication controller and route

* feat: remove unused imports

* chore: remove uneeeded auth controller

* feat: update seed and password min length

* chore: add supertest package to E2E tests

* test: add authenticate test

* chore: configure environment for integration test

* chore: add prisma generate for env test

* chore: update the e2e action

* feat: remove uneeded migrate execution

* chore: update github action to e2e test

* shore: reorganize actions

* chore: update actions

* chore: update actions

* chore: remove coverage from e2e

* chore: update actions

* feat: remove uneeded custom env for vitest

* Users controllers (#16)

* feat: add create user controller

* feat: fix returning from user create controller and add test

* feat: update auth test to add the create controller

* feat: update use case variable name from create user controller

* feat: add update user controller and update auth middleware

* chore: rename test case

* feat: add controller to delete user; update create and update controllers;

* feat: add test to update user and update another tests

* chore: just update test emails to keep most readable

* feat: Add controller to get profile user info

* chore: update vitest e2e test config to run at single thread

* feat: remove unused imports

* chore: update vite config to remove fileParallelism

* feat: refact pages controller (#17)

* feat: refact pages controller

* feat: update tests cases

* feat: add new use case and cnotroller to get links from specific page

* feat: refact links controllers and create tests cases (#18)

* chore: update readme with new specs

+++ src/use-cases/links/delete-link.ts
@@ -1,9 +1,16 @@
 import { LinksRepository } from '@/repositories/links-repository';
+import { ResourceNotFoundError } from '../errors/not-found-error';
 
 export class DeleteLinkUseCase {
   constructor(private linksRepository: LinksRepository) {}
 
   async execute(id: string) {
+    const link = await this.linksRepository.findById(id);
+
+    if (!link) {
+      throw new ResourceNotFoundError('Link not found');
+    }
+
     await this.linksRepository.delete(id);
   }
 }
